frontend_ttkb/requirements.txt
# Dependencias principales
ttkbootstrap>=1.10.1
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0  # Para manejo de imágenes en tkinter

# Utilidades para desarrollo
black>=23.0.0
mypy>=1.0.0
